feat: install and integrate lazygit (and related)
Kind of a tricky one to understand, in the way that what "integrate into
a workflow" means. For now, best I could think of is getting familiar
with it, making an alias, adding a shortcut to open lg in VSCode, and
setting my default git editor to VSCode.

I'll need to see how it goes later on to actually make a proper choice.